+++ rumqttc/CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+* Make v4::Connect::write return correct value
+
 ### Security
 
 ---